Gjør om felt totaltFeilutbetaltBeløp i hendelse fra tilbakekreving om til BigDecimal

+++ src/main/kotlin/no/nav/tilleggsstonader/sak/tilbakekreving/domene/Tilbakekrevingstatus.kt
@@ -2,6 +2,7 @@ package no.nav.tilleggsstonader.sak.tilbakekreving.domene
 
 import com.fasterxml.jackson.annotation.JsonSubTypes
 import com.fasterxml.jackson.annotation.JsonTypeInfo
+import java.math.BigDecimal
 import java.time.LocalDate
 import java.time.LocalDateTime
 
@@ -19,7 +20,7 @@ data class Tilbakekrevingsstatus(
     val sakOpprettet: LocalDateTime,
     val varselSendtTidspunkt: LocalDateTime?,
     val behandlingstatus: String,
-    val totaltFeilutbetaltBeløp: Long,
+    val totaltFeilutbetaltBeløp: BigDecimal,
     val tilbakekrevingFom: LocalDate,
     val tilbakekrevingTom: LocalDate,
     val tilbakekrevingBehandlingId: String,
